Fetch apt keys over https
This is the initial point of trust, everything else can be over
http (and verified with the keys), but if the keys are fetched over
http then they can potentially be modified in transit and so can't
be trusted.

Fetching over https, particularly from keyserver.ubuntu.com is a
sufficient initial level of trust for the keys.